Allow a user to disable oci-register-machine
oci-register-machine can be considered an information leak for
certain container workloads.  It allows information about the container
to be viewed by non privileged users.  Some adinistrators may want to disable
containers reporting this information to systemd-machinectl.  We want this
package installed by default, and removing this package is not an option
on atomic host.

This patch adds a /etc/oci-register-machine.conf which contains a json
data to disable the tool.